requirements.txt
selenium
pytest
webdriver-manager
mysql-connector-python
